requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
google-generativeai